Update design for Spacewalk Issue and Redeeem  (#253)
* extract base Form component

* separate issue and redeem in folders and use new From input

* extract feebox and dialog from issue

* issue box restyling with collapsible boxes

* fix collapsible styles

* tweak asset selector to match designs

* move coin icons to shared

* mock disclaimer text

* remove typo

* add stellar logo for tab

* fix pen/ampe logos and colors

* fixes

* settings dialog (wip)

* add 'back to stellar' design

* yarn lint fix

* Add settings Dialog

* improve from component

* improve settings

* Update src/components/Layout/links.tsx

* rename hook

* remove unused spacewalk app

* fix tricky dropdown issue not closing on blur

* improve settings dialog

* improve collapsible behavior

* make tabs slightly bigger

* invert buttons order

* remove minimum

* extract useBalances hook

* extract get asset display name

* fix redeem validation and max tokens

* make balance reload when balances hook is ready

* max issuable tokens

* fix typo

* fix showing errors

* add missing dot

* remove unused code

* show errors according to the design
