+++ src/assets/AmplitudeLogo.tsx
@@ -5,31 +5,18 @@ interface Props extends HTMLAttributes<SVGSVGElement> {
 }
 
 const AmplitudeLogo = ({ className, ...rest }: Props) => (
-  <svg fill="none" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" className={className} {...rest}>
-    <path
-      d="M3.8606 9.50286C3.8606 9.50286 4.34281 10.9582 4.61302 11.6738C4.88323 12.3895 5.03867 12.8273 5.71325 12.8259C5.99852 12.8252 6.29641 12.4543 6.62941 11.6738C7.01913 10.6583 7.4623 9.50286 7.4623 9.50286H3.8606Z"
-      fill="#4EE59A"
-    />
-    <path
-      d="M7.46228 9.50286C7.46228 9.50286 6.98006 8.04752 6.70985 7.33188C6.43964 6.61625 6.28354 6.17847 5.60953 6.18002C5.32426 6.18067 5.02638 6.55167 4.69338 7.33205C4.30366 8.34748 3.86057 9.50286 3.86057 9.50286H7.46228Z"
-      fill="#4EE59A"
-    />
-    <path
-      d="M15.2085 9.50286C15.2085 9.50286 13.2849 5.911 12.997 5.33482C12.7092 4.75865 12.2485 3.93145 11.3108 3.93145C10.3731 3.93145 9.9888 4.83443 9.75093 5.33482C9.51305 5.83521 7.46237 9.50286 7.46237 9.50286L15.2085 9.50286Z"
-      fill="#4EE59A"
-    />
-    <path
-      d="M7.4624 9.50286C7.4624 9.50286 9.38604 13.0947 9.67385 13.6709C9.96166 14.2471 10.4224 15.0743 11.3601 15.0743C12.2978 15.0743 12.6821 14.1713 12.92 13.6709C13.1578 13.1705 15.2085 9.50286 15.2085 9.50286H7.4624Z"
-      fill="#4EE59A"
-    />
+  <svg fill="#4EE59A" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" className={className} {...rest}>
+    <path d="M3.8606 9.50286C3.8606 9.50286 4.34281 10.9582 4.61302 11.6738C4.88323 12.3895 5.03867 12.8273 5.71325 12.8259C5.99852 12.8252 6.29641 12.4543 6.62941 11.6738C7.01913 10.6583 7.4623 9.50286 7.4623 9.50286H3.8606Z" />
+    <path d="M7.46228 9.50286C7.46228 9.50286 6.98006 8.04752 6.70985 7.33188C6.43964 6.61625 6.28354 6.17847 5.60953 6.18002C5.32426 6.18067 5.02638 6.55167 4.69338 7.33205C4.30366 8.34748 3.86057 9.50286 3.86057 9.50286H7.46228Z" />
+    <path d="M15.2085 9.50286C15.2085 9.50286 13.2849 5.911 12.997 5.33482C12.7092 4.75865 12.2485 3.93145 11.3108 3.93145C10.3731 3.93145 9.9888 4.83443 9.75093 5.33482C9.51305 5.83521 7.46237 9.50286 7.46237 9.50286L15.2085 9.50286Z" />
+    <path d="M7.4624 9.50286C7.4624 9.50286 9.38604 13.0947 9.67385 13.6709C9.96166 14.2471 10.4224 15.0743 11.3601 15.0743C12.2978 15.0743 12.6821 14.1713 12.92 13.6709C13.1578 13.1705 15.2085 9.50286 15.2085 9.50286H7.4624Z" />
     <path
       fillRule="evenodd"
       clipRule="evenodd"
       d="M9.5 0C11.3789 0 13.2156 0.557072 14.7779 1.60077C16.3402 2.64447 17.5578 4.12792 18.2769 5.86353C18.9959 7.59914 19.184 9.50895 18.8175 11.3515C18.4509 13.194 17.5461 14.8864 16.2175 16.2148C14.8889 17.5432 13.1962 18.4478 11.3534 18.8143C9.51054 19.1808 7.6004 18.9927 5.86451 18.2738C4.12861 17.5549 2.64491 16.3375 1.60104 14.7755C0.557165 13.2134 0 11.377 0 9.49841C0 6.97928 1.00089 4.56332 2.78249 2.78202C4.56408 1.00072 6.98044 0 9.5 0ZM9.5 1.86439C7.98986 1.86439 6.51364 2.31212 5.25801 3.15097C4.00238 3.98982 3.02373 5.1821 2.44583 6.57705C1.86793 7.972 1.71672 9.50697 2.01133 10.9878C2.30595 12.4687 3.03314 13.829 4.10097 14.8966C5.1688 15.9643 6.52929 16.6914 8.01041 16.9859C9.49153 17.2805 11.0267 17.1293 12.4219 16.5515C13.8171 15.9737 15.0096 14.9952 15.8486 13.7398C16.6876 12.4844 17.1354 11.0084 17.1354 9.4985C17.1354 7.47381 16.3309 5.53204 14.899 4.10037C13.4671 2.6687 11.525 1.86439 9.5 1.86439Z"
-      fill="#4EE59A"
     />
-    <path d="M7.50378 9.51406H15.0958" stroke="#4EE59A" />
-    <path d="M3.90198 9.51406H7.43313" stroke="#4EE59A" />
+    <path d="M7.50378 9.51406H15.0958" />
+    <path d="M3.90198 9.51406H7.43313" />
   </svg>
 );
