+++ src/components/Form/Validation/index.tsx
@@ -9,7 +9,7 @@ const Validation = ({ errors = {}, className }: ValidationProps): JSX.Element | 
   const keys = Object.keys(errors);
   if (keys.length === 0) return null;
   return (
-    <ul className={`text-red-700 text-sm ${className}`}>
+    <ul className={`text-red-400 text-sm ${className}`}>
       {keys.map((key, i) => (errors[key] && errors[key]?.message ? <li key={i}>{errors[key]?.message}</li> : null))}
     </ul>
   );
