+++ src/components/Layout/links.tsx
@@ -54,7 +54,7 @@ export const links: Links = ({ tenantName }) => [
     link: './spacewalk',
     title: 'Spacewalk',
     props: {
-      className: ({ isActive } = {}) => (isActive ? 'active' : tenantName === TenantName.Pendulum ? 'coming-soon' : ''),
+      className: ({ isActive } = {}) => (isActive ? 'active' : tenantName === TenantName.Pendulum ? 'active' : ''),
     },
     prefix: <SpacewalkIcon />,
     suffix: tenantName === TenantName.Pendulum ? <ComingSoonTag /> : <></>,
@@ -96,8 +96,6 @@ export const links: Links = ({ tenantName }) => [
     link: '/nabla',
     title: 'Nabla',
     prefix: <NablaIcon />,
-    suffix: <ComingSoonTag />,
-    hidden: true,
     props: {
       className: ({ isActive } = {}) => (isActive ? 'active' : ''),
     },
