+++ src/config/apps/index.ts
@@ -2,8 +2,7 @@
 import { nablaConfig } from './nabla';
 import { AppConfig } from './types';
 
-export const apps = 'nabla';
-export type Apps = typeof apps;
+export type Apps = 'nabla';
 
 export const appsConfigs = {
   nabla: nablaConfig,
