+++ src/helpers/yup.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export const transformNumber = (value: any, originalValue: any) => {
-  if (!originalValue) return value;
-  if (typeof originalValue === 'string') value = Number(originalValue) ?? 0;
+  if (!originalValue) return 0;
+  if (typeof originalValue === 'string' && originalValue !== '') value = Number(originalValue) ?? 0;
   return value;
 };
