+++ src/pages/bridge/styles.css
@@ -80,11 +80,63 @@
   fill: var(--table-text);
 }
 
-.card.bridge-card {
-  background-color: rgba(255, 255, 255, 0.06);
+
+.bridge-card .tabs-boxed .tab-active:not(.tab-disabled):not([disabled]) > svg {
+  fill: #32253E;
 }
 
-.bridge-card .tabs-boxed .tab-active:not(.tab-disabled):not([disabled]),
-.bridge-card button.btn-primary {
-  color: #202020;
+.bridge-card .tabs-boxed .tab-active:not(.tab-disabled):not([disabled]) {
+  color: #32253E;
+}
+
+.bridge-card input[type='number'] {
+  outline: 0px;
+}
+
+[data-theme='pendulum'] {
+  .bridge-card .tabs-boxed :not(.tab-active) > svg{
+    fill: var(--gray-3, #828282);
+  }
+
+  .bridge-card .tabs-boxed :not(.tab-active) {
+    color: var(--gray-3, #828282);
+  }
+
+  .bridge-card .tabs-boxed .tab-active:not(.tab-disabled):not([disabled]) {
+    background-color: #f7f7f7;
+  }
+
+  .bridge-card .disclaimer svg{
+    fill: #32253E
+  }
+
+  .bridge-card .settings svg {
+    stroke: #58667E;
+  }
+}
+
+[data-theme='amplitude'] {
+  .bridge-card .tabs-boxed :not(.tab-active) > svg{
+    fill: var(--primary);
+  }
+
+  .bridge-card .tabs-boxed :not(.tab-active) {
+    color: #fff;
+  }
+
+  .bridge-card .tabs-boxed .tab-active:not(.tab-disabled):not([disabled]) {
+    background-color: var(--primary);
+  }
+
+  .bridge-card .disclaimer svg {
+    fill: #fff;
+  }
+
+  .bridge-card input[type='number'] {
+    color: #fff;
+  }  
+
+  .bridge-card .settings svg {
+    stroke: #fff;
+  }
 }
