+++ tailwind.config.cjs
@@ -34,7 +34,8 @@ module.exports = {
           neutral: '#EFF2F5',
           'base-100': '#F5F9FA',
           'base-200': '#fff',
-          'base-300': '#eff2f5',
+          'base-300': '#F7F7F7',
+          'base-400': '#E7E7E7',
           'base-content': '#58667E',
           '--rounded-btn': '9px',
           '--btn-text-case': 'none',
@@ -52,6 +53,7 @@ module.exports = {
           'base-100': '#202020',
           'base-200': '#1c1c1c',
           'base-300': '#2c2c2c',
+          'base-400': '#E7E7E7',
           'base-content': '#fff',
           '--rounded-btn': '9px',
           '--btn-text-case': 'none',
